requirements.txt
requests
pygame-ce
websocket-client
